refactor(web): move enqueue test helper to db-test-seeders to remove eslint-disable
